Extract stripHtmlAttributes to lib/editor/ using isomorphic-dompurify
Move stripHtmlAttributes to a shared utility using isomorphic-dompurify
so it works in both browser and Node.js. Update all importers.
